requirements.txt
anyio==3.6.2
appnope==0.1.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
attrs==22.1.0
backcall==0.2.0
bcrypt==4.0.1
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
debugpy==1.6.4
decorator==5.1.1
defusedxml==0.7.1
dnspython==2.2.1
email-validator==1.3.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.2
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-WTF==1.0.1
fqdn==1.5.1
idna==3.4
ipykernel==6.19.2
ipython==8.7.0
ipython-genutils==0.2.0
ipywidgets==8.0.3
isoduration==20.11.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-events==0.5.0
jupyter_client==7.4.8
jupyter_core==5.1.0
jupyter_server==2.0.1
jupyter_server_terminals==0.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.4
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
mistune==2.0.4
mysql-connector==2.2.9
mysql-connector-python==8.0.31
nbclassic==0.4.8
nbclient==0.7.2
nbconvert==7.2.6
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.2
notebook_shim==0.2.2
numpy==1.23.5
packaging==22.0
pandocfilters==1.5.0
parso==0.8.3
peewee==3.15.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.0
prometheus-client==0.15.0
prompt-toolkit==3.0.36
protobuf==3.20.1
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
pycparser==2.21
Pygments==2.13.0
PyJWT==2.6.0
PyMySQL==1.0.2
pyrsistent==0.19.2
python-dateutil==2.8.2
python-json-logger==2.0.4
PyYAML==6.0
pyzmq==24.0.1
qtconsole==5.4.0
QtPy==2.3.0
requests==2.28.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
Send2Trash==1.8.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
terminado==0.17.1
tinycss2==1.2.1
tornado==6.2
traitlets==5.7.1
uri-template==1.2.0
urllib3==1.26.13
wcwidth==0.2.5
webcolors==1.12
webencodings==0.5.1
websocket-client==1.4.2
Werkzeug==2.2.2
widgetsnbextension==4.0.4
WTForms==3.0.1
